+++ packages/core/src/components/date-input/date-input.tsx
@@ -24,7 +24,11 @@ import {
 import { DateTime } from 'luxon';
 import { dropdownController } from '../dropdown/dropdown-controller';
 import { SlotEnd, SlotStart } from '../input/input.fc';
-import { adjustPaddingForStartAndEnd } from '../input/input.util';
+import {
+  DisposableChangesAndVisibilityObservers,
+  addDisposableChangesAndVisibilityObservers,
+  adjustPaddingForStartAndEnd,
+} from '../input/input.util';
 import {
   ClassMutationObserver,
   HookValidationLifecycle,
@@ -42,6 +46,8 @@ export type DateInputValidityState = {
 /**
  * @since 2.6.0
  * @form-ready 2.6.0
+ * @slot start - Element will be displayed at the start of the input
+ * @slot end - Element will be displayed at the end of the input
  */
 @Component({
   tag: 'ix-date-input',
@@ -172,6 +178,8 @@ export class DateInput implements IxInputFieldComponent<string> {
   private classObserver?: ClassMutationObserver;
   private invalidReason?: string;
 
+  private disposableChangesAndVisibilityObservers?: DisposableChangesAndVisibilityObservers;
+
   updateFormInternalValue(value: string): void {
     this.formInternals.setFormValue(value);
     this.value = value;
@@ -181,6 +189,12 @@ export class DateInput implements IxInputFieldComponent<string> {
     this.classObserver = createClassMutationObserver(this.hostElement, () =>
       this.checkClassList()
     );
+
+    this.disposableChangesAndVisibilityObservers =
+      addDisposableChangesAndVisibilityObservers(
+        this.hostElement,
+        this.updatePaddings.bind(this)
+      );
   }
 
   componentWillLoad(): void {
@@ -195,10 +209,6 @@ export class DateInput implements IxInputFieldComponent<string> {
     this.updateFormInternalValue(this.value);
   }
 
-  componentDidRender(): void {
-    this.updatePaddings();
-  }
-
   private updatePaddings() {
     adjustPaddingForStartAndEnd(
       this.slotStartRef.current,
@@ -208,9 +218,8 @@ export class DateInput implements IxInputFieldComponent<string> {
   }
 
   disconnectedCallback(): void {
-    if (this.classObserver) {
-      this.classObserver.destroy();
-    }
+    this.classObserver?.destroy();
+    this.disposableChangesAndVisibilityObservers?.();
   }
 
   @Watch('value')