+++ packages/react-test-app/src/main.tsx
@@ -218,6 +218,10 @@ import LayoutAutoCustom from './preview-examples/layout-auto-custom';
 import FormCheckboxGroupIndeterminate from './preview-examples/form-checkbox-group-indeterminate';
 import FormLayoutAuto from './preview-examples/form-layout-auto';
 import FormLayoutGrid from './preview-examples/form-layout-grid';
+import InputWithSlots from './preview-examples/input-with-slots';
+import NumberInputWithSlots from './preview-examples/number-input-with-slots';
+import DateInput from './preview-examples/date-input';
+import DateInputWithSlots from './preview-examples/date-input-with-slots';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <IxApplicationContext>
@@ -600,6 +604,7 @@ ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
         <Route path="/preview/input-readonly" element={<InputReadonly />} />
         <Route path="/preview/input-types" element={<InputTypes />} />
         <Route path="/preview/input-validation" element={<InputValidation />} />
+        <Route path="/preview/input-with-slots" element={<InputWithSlots />} />
         <Route path="/preview/textarea-legacy" element={<TextareaLegacy />} />
         <Route path="/preview/theme-switcher" element={<ThemeSwitcher />} />
         <Route path="/preview/tile" element={<Tile />} />
@@ -642,6 +647,10 @@ ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
         <Route path="/preview/tree-custom" element={<TreeCustom />} />
         <Route path="/preview/number-input" element={<NumberInput />} />
         <Route
+          path="/preview/number-input-with-slots"
+          element={<NumberInputWithSlots />}
+        />
+        <Route
           path="/preview/number-input-disabled"
           element={<NumberInputDisabled />}
         />
@@ -660,6 +669,11 @@ ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
         <Route
           path="/preview/number-input-validation"
           element={<NumberInputValidation />}
+        />
+        <Route path="/preview/date-input" element={<DateInput />} />
+        <Route
+          path="/preview/date-input-with-slots"
+          element={<DateInputWithSlots />}
         />
         <Route path="/preview/tree" element={<Tree />} />
         <Route path="/preview/upload" element={<Upload />} />