+++ packages/vue/src/components.ts
@@ -132,7 +132,8 @@ export const IxApplication = /*@__PURE__*/ defineContainer<JSX.IxApplication>('i
 export const IxApplicationHeader = /*@__PURE__*/ defineContainer<JSX.IxApplicationHeader>('ix-application-header', defineIxApplicationHeader, [
   'name',
   'showMenu',
-  'menuToggle'
+  'menuToggle',
+  'openAppSwitch'
 ]);
 
 
@@ -332,6 +333,10 @@ export const IxDateDropdown = /*@__PURE__*/ defineContainer<JSX.IxDateDropdown>(
   'minDate',
   'maxDate',
   'dateRangeId',
+  'variant',
+  'outline',
+  'ghost',
+  'loading',
   'customRangeAllowed',
   'dateRangeOptions',
   'locale',
@@ -516,6 +521,9 @@ export const IxExpandingSearch = /*@__PURE__*/ defineContainer<JSX.IxExpandingSe
   'placeholder',
   'value',
   'fullWidth',
+  'variant',
+  'outline',
+  'ghost',
   'valueChange'
 ]);
 
@@ -722,7 +730,8 @@ export const IxMenu = /*@__PURE__*/ defineContainer<JSX.IxMenu>('ix-menu', defin
   'i18nExpand',
   'i18nCollapse',
   'expandChange',
-  'mapExpandChange'
+  'mapExpandChange',
+  'openAppSwitch'
 ]);
 
 
@@ -977,6 +986,8 @@ export const IxSelect = /*@__PURE__*/ defineContainer<JSX.IxSelect, JSX.IxSelect
   'i18nSelectListHeader',
   'i18nNoMatches',
   'hideListHeader',
+  'dropdownWidth',
+  'dropdownMaxWidth',
   'valueChange',
   'itemSelectionChange',
   'inputChange',